cursor/backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
PyMySQL==1.1.1
dj-database-url==2.2.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
